Re-architecture Phase 4: backend SPI and capability model (backend.api.spi)
Adds the SPI types decided in the design review: Backend (deployment-scoped:
type, capabilities, newSession, close), BackendSession (the domain-API
accessors bound to a user; contract javadoc pins cheap per-request creation,
no object-identity guarantees, capability-gated accessors, and null-userId
workspace specification resolution), BackendSessionContext with
BackendSessionStateStore (identity crosses as data; per-user backend state -
e.g. OAuth tokens - crosses as a host-implemented string-keyed store),
BackendFactory, BackendConfiguration (Jackson-polymorphic base, type
discriminator with inline subtype fields), BackendEnvironment (object mapper,
task processor, and the deployment's structure extension provider + platform
extensions passed through to generic code, never bundled in a backend jar),
the BackendCapability starter enum, UnsupportedCapabilityException (extends
LegendSDLCException with status 501, so the existing mapper already produces
the decided status; carries capability and backend type), and
AuthorizationRequiredException (authorization URI as data).

The session's auth-flow surface (authorize/callback/terms-of-service) is
deliberately deferred to the GitLab backend step of this phase so it is
shaped against the real flows. BackgroundTaskProcessor relocates from
server.tools to org.finos.legend.sdlc.backend.api.tools (deprecated
constructor-forwarding bridge at the old FQN); it cannot live in
legend-sdlc-shared, which has no compile dependencies, because it needs
slf4j. Full-reactor mvn install javadoc:javadoc green.
