+++ legend-sdlc-backend-api/src/main/java/org/finos/legend/sdlc/backend/api/spi/BackendConfiguration.java
@@ -1,0 +1,28 @@
+// Copyright 2026 Goldman Sachs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.finos.legend.sdlc.backend.api.spi;
+
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+
+/**
+ * Base of a backend's configuration, resolved polymorphically from the host configuration's {@code backend:}
+ * section by the {@code type} discriminator, with the subtype's own fields inline at the same level. The host
+ * registers each discovered {@link BackendFactory}'s {@link BackendFactory#getConfigurationClass() configuration
+ * class} as a subtype named by {@link BackendFactory#getType()}.
+ */
+@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
+public abstract class BackendConfiguration
+{
+}
