+++ legend-sdlc-server/src/main/java/org/finos/legend/sdlc/server/gitlab/api/AbstractGitlabWorkflowApi.java
@@ -20,7 +20,7 @@ import org.eclipse.collections.impl.factory.primitive.LongObjectMaps;
 import org.finos.legend.sdlc.server.gitlab.GitLabConfiguration;
 import org.finos.legend.sdlc.server.gitlab.auth.GitLabUserContext;
 import org.finos.legend.sdlc.server.gitlab.tools.PagerTools;
-import org.finos.legend.sdlc.server.tools.BackgroundTaskProcessor;
+import org.finos.legend.sdlc.backend.api.tools.BackgroundTaskProcessor;
 import org.gitlab4j.api.GitLabApiException;
 import org.gitlab4j.api.MergeRequestApi;
 import org.gitlab4j.api.Pager;
